requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==4.0.3
attrs==26.1.0
bcrypt==5.0.0
blinker==1.9.0
build==1.4.3
cachetools==6.2.6
certifi==2026.2.25
charset-normalizer==3.4.7
chromadb==1.5.7
click==8.1.8
coloredlogs==15.0.1
dataclasses-json==0.6.7
distro==1.9.0
durationpy==0.10
exceptiongroup==1.3.1
filelock==3.19.1
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.46
googleapis-common-protos==1.74.0
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.8.0
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.14.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
langchain==0.3.28
langchain-community==0.3.31
langchain-core==0.3.84
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langsmith==0.4.37
markdown-it-py==3.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
narwhals==2.19.0
numpy==2.0.2
oauthlib==3.3.1
onnxruntime==1.19.2
openai==2.31.0
opentelemetry-api==1.41.0
opentelemetry-exporter-otlp-proto-common==1.41.0
opentelemetry-exporter-otlp-proto-grpc==1.41.0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
orjson==3.11.5
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
propcache==0.4.1
protobuf==6.33.6
pyarrow==21.0.0
pybase64==1.4.3
pydantic==2.13.1
pydantic-settings==2.11.0
pydantic_core==2.46.1
pydeck==0.9.1
Pygments==2.20.0
pypdf==6.10.2
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.36.2
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==15.0.0
rpds-py==0.27.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
SQLAlchemy==2.0.49
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
tomli==2.4.1
tornado==6.5.5
tqdm==4.67.3
typer==0.23.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.39.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
yarl==1.22.0
zipp==3.23.1
zstandard==0.25.0
